feat(dwm): Add transparency and modify tag setup
This commit introduces transparency to dwm's status bar and makes some
modifications to the tag setup and terminal emulator.

- Add alpha patch for translucent status bar
- Reduce number of tags from 9 to 3
- Change default terminal from st to Alacritty

Changes in detail:
- Introduce baralpha and borderalpha settings in config.h
- Add alphas array for configuring transparency levels
- Modify tags array to only include "1", "2", "3"
- Update termcmd to use "alacritty" instead of "st"
- Add dwm-alpha-20230401-348f655.diff patch
- Update patch_order.txt to include the new alpha patch

This update enhances the visual appeal of dwm with a translucent bar
while simplifying the tag setup for a more focused workflow. The switch
to Alacritty offers a modern terminal experience.